add resource list api ### PR types feature
### PR changes
feature

### Describe
add resource list api
